# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Board API'
  description: 'HTTP JSON API'
  contact:
    name: 'Laravel api'
  license:
    name: Mr.Dow
  version: 1.0.0
servers:
  -
    url: /api
    description: 'API server'
paths:
...
  '/user/adverts/{advertId}':
    get:
      tags:
        - Adverts
      operationId: 'App\Http\Controllers\Api\User\AdvertController::show'
      parameters:
        -
          name: advertId
          in: path
          description: 'ID of advert'
          required: true
      responses:
        '200':
          description: 'Success response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvertShowResource'
      security:
        -
          Bearer: []
          OAuth2: []
    put:
      tags:
        - Adverts
      operationId: 'App\Http\Controllers\Api\User\AdvertController::update'
      parameters:
        -
          name: advertId
          in: path
          description: 'ID of advert'
          required: true
      requestBody:
        description: 'update user''s advert'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRequest'
      responses:
        '201':
          description: 'Success response'
      security:
        -
          Bearer: []
          OAuth2: []
    delete:
      tags:
        - Adverts
      operationId: 'App\Http\Controllers\Api\User\AdvertController::destroy'
      parameters:
        -
          name: advertId
          in: path
          required: true
      responses:
        '204':
          description: 'Success response'
      security:
        -
          Bearer: []
          OAuth2: []
...
components:
  schemas:
...
    EditRequest:
      properties:
        title:
          type: string
        content:
          type: string
        price:
          type: string
        address:
          type: string
      type: aplication/json
...
    AdvertShowResource:
      properties:
        id:
          type: integer
        user_name:
          type: string
        phone:
          type: string
        category:
          type: integer
        region:
          type: integer
        title:
          type: string
        content:
          type: string
        price:
          type: string
        address:
          type: string
        published:
          type: string
        expires:
          type: string
      type: application/json
...
  securitySchemes:
    OAuth2:
      type: oauth2
      description: 'Use a global client_id / client_secret and your username / password combo to obtain a token'
      flows:
        password:
          tokenUrl: 'https://localhost:8080/oauth/token'
          scopes:
            description: 'You can use all functions in this site'
    Bearer:
      type: apiKey
      name: Authorization
      in: header